+++ oppia/viz/views.py
@@ -40,25 +40,27 @@ def summary_view(request):
         else:
             other_country_activity += c['country_total_hits']
         i += 1
-    if i >= 20:
+    if i > 20:
         hits_percent = float(other_country_activity * 100.0/total_hits['total_hits'])
         country_activity.append({'country_code':None,'country_name':_('Other'),'hits_percent':hits_percent })
         
     # Language
-    hit_by_language = Tracker.objects.filter(user__is_staff=False).values('lang').annotate(total_hits=Count('id')).order_by('-total_hits')
-    total_hits = Tracker.objects.filter(user__is_staff=False).aggregate(total_hits=Count('id'))
+    hit_by_language = Tracker.objects.filter(user__is_staff=False).exclude(lang=None).values('lang').annotate(total_hits=Count('id')).order_by('-total_hits')
+    total_hits = Tracker.objects.filter(user__is_staff=False).exclude(lang=None).aggregate(total_hits=Count('id'))
     
     i = 0
     languages = []
     other_languages = 0
     for hbl in hit_by_language:
+        print hbl['lang']
+        print hbl['total_hits']
         if i < 10:
             hits_percent = float(hbl['total_hits'] * 100.0/total_hits['total_hits'])
             languages.append({'lang':hbl['lang'],'hits_percent':hits_percent })
         else:
-            other_languages += hbc['total_hits']
+            other_languages += hbl['total_hits']
         i += 1
-    if i >= 10:
+    if i > 10:
         hits_percent = float(other_languages * 100.0/total_hits['total_hits'])
         languages.append({'lang':_('Other'),'hits_percent':hits_percent })
         
@@ -91,7 +93,7 @@ def summary_view(request):
         else:
             other_course_activity += hbc['total_hits']
         i += 1
-    if i >= 10:
+    if i > 10:
         hits_percent = float(other_course_activity * 100.0/total_hits['total_hits'])
         hot_courses.append({'course':_('Other'),'hits_percent':hits_percent })
                        
@@ -99,6 +101,7 @@ def summary_view(request):
                               {'user_registrations': user_registrations,
                                'total_countries': total_countries,
                                'country_activity': country_activity,
+                               'languages': languages,
                                'course_downloads': course_downloads,
                                'course_activity': course_activity, 
                                'hot_courses': hot_courses, },