+++ docs/changelog.rst
@@ -4,6 +4,7 @@ Change Log
 v0.2.14
 -------
 * Fix https://github.com/DigitalCampus/django-oppia/issues/133 - add visualizations page
+* Fix https://github.com/DigitalCampus/django-oppia/issues/137 - add lang to tracker model
 
 v0.2.12/13
 ----------